added method to get number of bytes the client reported to have received
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@2434 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/net/rtmp/RTMPConnection.java
@@ -173,7 +173,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 	/**
 	 * Number of bytes the client reported to have received.
 	 */
-	private int clientBytesRead = 0;
+	private long clientBytesRead = 0;
 
 	/**
 	 * Bandwidth configure.
@@ -639,7 +639,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 	 * 
 	 * @return Number of bytes
 	 */
-	public int getClientBytesRead() {
+	public long getClientBytesRead() {
 		return clientBytesRead;
 	}
